Fix ESLint errors and improve typing
