+++ src/app/core/_components/tables/access-permission-groups-user-table/access-permission-groups-user-table.component.html
@@ -11,6 +11,6 @@
   [isFilterable]="true"
   [filterFn]="filter"
   [isPageable]="true"
-  (editableCheckbox)="editableCheckSaved($event)"
+  (editableCheckbox)="onCheckboxChange($event)"
   (exportActionClicked)="exportActionClicked($event)"
 />
