--- src/app/core/_models/user-permissions.model.ts
@@ -1,8 +1,0 @@
-export interface UserPermissions {
-  name: string;
-  key: string;
-  create: boolean;
-  read: boolean;
-  update: boolean;
-  delete: boolean;
-}
